btrfs/172: cleanup logwrites on failure
This requires a specific mkfs option, but it's incompatible with newer
mkfs options, so mkfs can fail when using those options.  When this
happens the logwrites device is left which causes failures in the rest
of the suite.  Add a _cleanup() to make sure the logwrites device is
cleaned up.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>